map -> nore, migrate imaps to <plug>
This fixes the offensive imaps. For example, consider letting mapleader
= ' '. Then every time you press space in insert mode, this plugin will
freeze the input until either the timeout for mappings is reached, or it
will freeze indefinitely. It is also impossible with such a
configuration to type " is" in insert mode without escaping to normal
mode after the first 'i'. This commit fixes this unbelievably annoying
oversight.

It also reconfigures the nmaps to be nnoremaps so as to avoid any other
(rarer) annoying behavior.